requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic[email]
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
fastapi-mail
slowapi
cloudinary
redis